Reduce reporting of DTWAIN_TN_GENERALERROR notifications to functions registered with DTWAIN_SetCallback and DTWAIN_SetCallback64.

+++ source/cpp/ctldib32.cpp
@@ -849,7 +849,7 @@ HANDLE CTL_TwainDib::CreateBMPBitmapFromDIB(HANDLE hDib)
         lpbi->biSize + CDibInterface::CalculateUsedPaletteEntries(bpp) * sizeof(RGBQUAD);
 
     // we need to attach the bitmap header info onto the data
-    const unsigned int totalSize = ImageMemoryHandler::GlobalSize(hDib) + sizeof(BITMAPFILEHEADER);
+    const size_t totalSize = ImageMemoryHandler::GlobalSize(hDib) + sizeof(BITMAPFILEHEADER);
 
     // Allocate for returned handle
     returnHandle = static_cast<HANDLE>(ImageMemoryHandler::GlobalAlloc(GMEM_FIXED, totalSize));
